--- document-legacy/src/error.rs
@@ -1,10 +1,0 @@
-use super::LayerId;
-
-/// A set of different errors that can occur when using this crate.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum DocumentError {
-	LayerNotFound(Vec<LayerId>),
-	InvalidPath,
-	NotFolder,
-	InvalidFile(String),
-}
