+++ document-legacy/src/layers/mod.rs
@@ -12,7 +12,6 @@
 //! When different layers overlap, they are blended together according to the [BlendMode](blend_mode::BlendMode)
 //! using the CSS [`mix-blend-mode`](https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode) property and the layer opacity.
 
-pub mod base64_serde;
 /// Contains the [FolderLegacyLayer](folder_layer::FolderLegacyLayer) type that encapsulates other layers, including more folders.
 pub mod folder_layer;
 /// Contains the base [Layer](layer_info::Layer) type, an abstraction over the different types of layers.
