+++ editor/src/messages/frontend/utility_types.rs
@@ -1,5 +1,5 @@
+use document_legacy::document::LayerId;
 use document_legacy::document_metadata::LayerNodeIdentifier;
-use document_legacy::LayerId;
 use serde::{Deserialize, Serialize};
 
 #[derive(PartialEq, Eq, Clone, Debug, Serialize, Deserialize, specta::Type)]
